+++ src/ExchangeRateProvider/BaseCurrencyProvider.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace Brick\Money\ExchangeRateProvider;
 
+use Brick\Math\BigInteger;
 use Brick\Math\BigNumber;
-use Brick\Math\BigRational;
+use Brick\Money\Currency;
+use Brick\Money\Exception\UnknownCurrencyException;
 use Brick\Money\ExchangeRateProvider;
 use Override;
 
@@ -20,32 +22,46 @@ use Override;
  */
 final readonly class BaseCurrencyProvider implements ExchangeRateProvider
 {
+    private Currency $baseCurrency;
+
     /**
-     * @param ExchangeRateProvider $provider         The provider for rates relative to the base currency.
-     * @param string               $baseCurrencyCode The code of the currency all the exchanges rates are based on.
+     * @param ExchangeRateProvider $provider     The provider for rates relative to the base currency.
+     * @param Currency|string      $baseCurrency The currency or ISO currency code all the exchanges rates are based on.
+     *
+     * @throws UnknownCurrencyException If an unknown currency code is given.
      */
     public function __construct(
         private ExchangeRateProvider $provider,
-        private string $baseCurrencyCode,
+        Currency|string $baseCurrency,
     ) {
+        $this->baseCurrency = $baseCurrency instanceof Currency ? $baseCurrency : Currency::of($baseCurrency);
     }
 
     #[Override]
-    public function getExchangeRate(string $sourceCurrencyCode, string $targetCurrencyCode): BigNumber
+    public function getExchangeRate(Currency $sourceCurrency, Currency $targetCurrency): ?BigNumber
     {
-        if ($sourceCurrencyCode === $this->baseCurrencyCode) {
-            return BigNumber::of($this->provider->getExchangeRate($sourceCurrencyCode, $targetCurrencyCode));
+        if ($sourceCurrency->isEqualTo($targetCurrency)) {
+            return BigInteger::one();
         }
 
-        if ($targetCurrencyCode === $this->baseCurrencyCode) {
-            $exchangeRate = $this->provider->getExchangeRate($targetCurrencyCode, $sourceCurrencyCode);
-
-            return BigRational::of($exchangeRate)->reciprocal();
+        if ($sourceCurrency->isEqualTo($this->baseCurrency)) {
+            return $this->provider->getExchangeRate($sourceCurrency, $targetCurrency);
         }
 
-        $baseToSource = $this->provider->getExchangeRate($this->baseCurrencyCode, $sourceCurrencyCode);
-        $baseToTarget = $this->provider->getExchangeRate($this->baseCurrencyCode, $targetCurrencyCode);
+        if ($targetCurrency->isEqualTo($this->baseCurrency)) {
+            $exchangeRate = $this->provider->getExchangeRate($targetCurrency, $sourceCurrency);
 
-        return BigRational::of($baseToTarget)->dividedBy($baseToSource);
+            return $exchangeRate?->toBigRational()->reciprocal();
+        }
+
+        $baseToSource = $this->provider->getExchangeRate($this->baseCurrency, $sourceCurrency);
+
+        if ($baseToSource === null) {
+            return null;
+        }
+
+        $baseToTarget = $this->provider->getExchangeRate($this->baseCurrency, $targetCurrency);
+
+        return $baseToTarget?->toBigRational()->dividedBy($baseToSource);
     }
 }